+++ projects/helgoland/d3/src/lib/model/d3-highlight.ts
@@ -1,0 +1,9 @@
+export interface HighlightOutput {
+    timestamp: number;
+    ids: Array<HighlightValue>;
+}
+
+export interface HighlightValue {
+    timestamp: number;
+    value: number;
+}
